moved integration test utilities function to main code. Modified build.sbt accordingly
