+++ src/main/java/com/hurence/logisland/integration/testutils/Publisher.java
@@ -1,4 +1,4 @@
-package com.hurence.logisland.integration;
+package com.hurence.logisland.integration.testutils;
 
 import java.io.IOException;
