+++ src/main/java/com/hurence/logisland/integration/testutils/SmallFileUtil.java
@@ -1,4 +1,4 @@
-package com.hurence.logisland.integration;
+package com.hurence.logisland.integration.testutils;
 
 import java.io.BufferedReader;
 import java.io.File;
