+++ src/test/java/com/hurence/logisland/integration/EmbeddedKafkaEnvironmentTest.java
@@ -1,5 +1,6 @@
 package com.hurence.logisland.integration;
 
+import com.hurence.logisland.integration.testutils.EmbeddedKafkaEnvironment;
 import org.junit.Test;
 
 import static org.junit.Assert.assertTrue;
